training updates

+++ src/text_summarizer/entities/config_entity.py
@@ -33,7 +33,7 @@ class ParamConfig:
     logging_steps: int
     evaluation_strategy: str
     eval_steps: int
-    save_steps: float
+    save_steps: int
     gradient_accumulation_steps: int
